removed compiler warnings and fiddled with build.xml to make cross-compiling for 64- and 32-bit easier
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@6001 23e1bc29-d954-0410-9766-47de34158ba0

+++ src/PLASMA/PlanDatabase/base/TokenType.cc
@@ -72,7 +72,7 @@ namespace EUROPA {
     }
 
     PSDataType* TokenType::getParameterType(int index) const {
-   	  check_error(index < m_args.size(), "Index out of bounds");
+      check_error((unsigned int) index < m_args.size(), "Index out of bounds");
    	  std::map<LabelStr,DataTypeId>::const_iterator it = m_args.begin();
    	  while (index-- > 0) ++it;
    	  return it->second;
